+++ plugins/panda/src/theme/recipes/qr-code.ts
@@ -1,0 +1,8 @@
+import { qrCodeAnatomy } from '@ark-ui/anatomy'
+import { defineSlotRecipe } from '@pandacss/dev'
+
+export const qrCode = defineSlotRecipe({
+  className: 'qrCode',
+  slots: qrCodeAnatomy.keys(),
+  base: {},
+})