+++ plugins/panda/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Fixed
+
+- Prepared recipes for `QrCode`, `SignaturePad` and `Field` components.
+
 ## [0.38.0] - 2024-06-06
 
 We are excited to announce that in this release, we have migrated all components to the new Ark UI 3.0.0 API.